api/requirements.txt
flask
flask-cors

google-api-python-client
google-auth
google-cloud
google-generativeai==0.8.5

python-dotenv
dotenv
pandas
xgboost
matplotlib
seaborn
requests
numpy==1.26.4
google-cloud-bigquery
scikit-learn
google-auth
google-auth-oauthlib
google-auth-httplib2
pmdarima


Flask-SQLAlchemy 
SQLAlchemy 
PyMySQL
dataframe_image
